pyproject.toml
[project]
name = "neostats-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.7.4",
    "langchain>=0.2,<1.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-groq>=0.1.0",
    "langchain-huggingface>=0.0.3",
    "langchain-openai>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "pydantic>=2.7,<3",
    "pypdf>=4.0.0",
    "sentence-transformers>=2.7.0",
    "streamlit>=1.35.0",
    "tavily-python>=0.3.0",
    "typing-extensions>=4.12",
    "python-dotenv>=1.0.1",
]
